+++ src/main/java/com/ducks/goodsduck/commons/controller/UserItemController.java
@@ -1,6 +1,6 @@
 package com.ducks.goodsduck.commons.controller;
 
-import com.ducks.goodsduck.commons.model.dto.ItemDto;
+import com.ducks.goodsduck.commons.model.dto.item.ItemDto;
 import com.ducks.goodsduck.commons.model.dto.LikeItemResponse;
 import com.ducks.goodsduck.commons.service.JwtService;
 import com.ducks.goodsduck.commons.service.UserItemService;
