+++ src/main/java/com/ducks/goodsduck/commons/model/dto/item/ItemDetailResponseCategory.java
@@ -1,4 +1,4 @@
-package com.ducks.goodsduck.commons.model.dto;
+package com.ducks.goodsduck.commons.model.dto.item;
 
 import com.ducks.goodsduck.commons.model.entity.CategoryItem;
 import com.ducks.goodsduck.commons.model.entity.Item;
