+++ src/main/java/com/ducks/goodsduck/commons/model/dto/item/ItemDetailResponseImage.java
@@ -1,4 +1,4 @@
-package com.ducks.goodsduck.commons.model.dto;
+package com.ducks.goodsduck.commons.model.dto.item;
 
 import com.ducks.goodsduck.commons.model.entity.Image;
 import lombok.Data;
