+++ src/main/java/com/ducks/goodsduck/commons/model/dto/oauth2/AuthorizationKakaoDto.java
@@ -1,14 +1,14 @@
-package com.ducks.goodsduck.commons.model.dto;
+package com.ducks.goodsduck.commons.model.dto.oauth2;
 
 import lombok.Data;
 
 @Data
-public class AuthorizationNaverDto {
+public class AuthorizationKakaoDto {
 
     private String access_token;
     private String refresh_token;
     private String token_type;
     private String expires_in;
     private String scope;
-    private String error_description;
+    private String refresh_token_expires_in;
 }
