+++ src/main/java/com/ducks/goodsduck/commons/model/dto/oauth2/AuthorizationNaverDto.java
@@ -1,14 +1,14 @@
-package com.ducks.goodsduck.commons.model.dto;
+package com.ducks.goodsduck.commons.model.dto.oauth2;
 
 import lombok.Data;
 
 @Data
-public class AuthorizationKakaoDto {
+public class AuthorizationNaverDto {
 
     private String access_token;
     private String refresh_token;
     private String token_type;
     private String expires_in;
     private String scope;
-    private String refresh_token_expires_in;
+    private String error_description;
 }
