+++ src/main/java/com/ducks/goodsduck/commons/model/dto/user/JwtDto.java
@@ -1,4 +1,4 @@
-package com.ducks.goodsduck.commons.model.dto;
+package com.ducks.goodsduck.commons.model.dto.user;
 
 import lombok.Data;
