+++ src/main/java/com/ducks/goodsduck/commons/model/entity/IdolGroup.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class IdolGroup {
 
-    @Id @GeneratedValue
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idol_group_id")
     private Long id;
     private String name;
