+++ src/main/java/com/ducks/goodsduck/commons/model/entity/IdolMember.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class IdolMember {
 
-    @Id @GeneratedValue
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idol_member_id")
     private Long id;
     private String name;
