+++ src/main/java/com/ducks/goodsduck/commons/model/enums/TradeStatus.java
@@ -1,5 +1,5 @@
 package com.ducks.goodsduck.commons.model.enums;
 
 public enum TradeStatus {
-    BUYING, FOR_SALE, TRADING, COMPLETE
+    BUYING, SELLING, TRADING, COMPLETE
 }
