+++ src/main/java/com/ducks/goodsduck/commons/repository/IdolMemberRepositoryCustom.java
@@ -8,6 +8,5 @@ import java.util.List;
 
 @Repository
 public interface IdolMemberRepositoryCustom {
-
     List<IdolMember> findAllByIdolGroupId(Long IdolGroupId);
 }
