+++ src/main/java/com/ducks/goodsduck/commons/service/CustomJwtService.java
@@ -1,6 +1,6 @@
 package com.ducks.goodsduck.commons.service;
 
-import com.ducks.goodsduck.commons.model.dto.JwtDto;
+import com.ducks.goodsduck.commons.model.dto.user.JwtDto;
 import com.ducks.goodsduck.commons.util.AwsSecretsManagerUtil;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
